+++ Shared/Models/Authentication/RegisterRequest.cs
@@ -1,12 +1,17 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Quibble.Shared.Api.Authentication
+namespace Quibble.Shared.Models.Authentication
 {
-    public class LoginRequest
+    public class RegisterRequest
     {
         [Required]
         [DataType("Username")]
         public string UserName { get; set; } = string.Empty;
+
+        [Required]
+        [EmailAddress]
+        [DataType(DataType.EmailAddress)]
+        public string Email { get; set; } = string.Empty;
 
         [Required]
         [DataType(DataType.Password)]
