+++ Shared/Shared.csproj
@@ -17,4 +17,5 @@
   <ItemGroup>
     <SupportedPlatform Include="browser" />
   </ItemGroup>
+  
 </Project>
